+++ packages/ui-components/src/icon-button/IconButton.tsx
@@ -15,7 +15,6 @@ function IconButton(props: IconButtonProps, ref: Ref<HTMLButtonElement>) {
     id,
     type = 'default',
     size = 'default',
-    active = 'false',
     disabled = false,
     loading = false,
     className,